+++ app/Console/Commands/FixRegionIssues.php
@@ -168,51 +168,4 @@ class FixRegionIssues extends Command
             ? $validator->errors()->first($fieldName)
             : null;
     }
-
-    /**
-     * Returns array of [organizationId => organizationIdentifier].
-     *
-     * @param $aidstreamOrganizationIds
-     *
-     * @return array
-     */
-    private function getAidstreamOrganizationIdentifier($aidstreamOrganizationIds): array
-    {
-        $returnArr = [];
-        $aidStreamSettings = $this->db::connection('aidstream')->table('settings')
-            ->whereIn('organization_id', $aidstreamOrganizationIds)
-            ->get();
-
-        if ($aidStreamSettings) {
-            foreach ($aidStreamSettings as $aidStreamSetting) {
-                if (in_array($aidStreamSetting->organization_id, $aidstreamOrganizationIds)) {
-                    $registryInfo = $aidStreamSetting->registry_info ? json_decode($aidStreamSetting->registry_info) : false;
-                    $organizationIdentifier = $registryInfo[0]?->publisher_id;
-                    $returnArr[$aidStreamSetting->organization_id] = $organizationIdentifier;
-                }
-            }
-        }
-
-        return $returnArr;
-    }
-
-    /**
-     * Returns mapped array of ids
-     * [aidstreamOrgId => iatiOrgId].
-     *
-     * @param array $aidstreamOrganizationIdentifierArray
-     * @param $iatiOrganizationIdArray
-     *
-     * @return array
-     */
-    private function mapOrganizationIds(array $aidstreamOrganizationIdentifierArray, $iatiOrganizationIdArray): array
-    {
-        $returnArr = [];
-
-        foreach ($aidstreamOrganizationIdentifierArray as $aidstreamId=>$identifier) {
-            $returnArr[$aidstreamId] = Arr::get($iatiOrganizationIdArray, $identifier, '');
-        }
-
-        return $returnArr;
-    }
 }
